Require 'types' directly from Sumsub module
Without doing this we receive an error if the gem is being used in a rails application and we try to require it, because 'types' is being used but it's not present in the Sumsub module.

I didn't realized that because I was testing the gem in `./bin/console` and there I was requiring 'types', what avoided the problem when I required 'sumsub'.